+++ lib/layrz_models.g.dart
@@ -983,8 +983,8 @@ Map<String, dynamic> _$$EmployeeImplToJson(_$EmployeeImpl instance) =>
     };
 
 const _$ColorblindModeEnumMap = {
-  ColorblindMode.propanopia: 'PROTANOPIA',
-  ColorblindMode.propanomaly: 'PROTANOMALY',
+  ColorblindMode.protanopia: 'PROTANOPIA',
+  ColorblindMode.protanomaly: 'PROTANOMALY',
   ColorblindMode.deuteranopia: 'DEUTERANOPIA',
   ColorblindMode.deuteranomaly: 'DEUTERANOMALY',
   ColorblindMode.tritanopia: 'TRITANOPIA',